refactor: rename utils/versioning to utils/tool_versioning to easily distinguish from domain/versioning
